Move hud-controller to its own file
